# lib/api-zod/src/generated/types/clusterSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSL AI Cricket Team Recommendation API
 * OpenAPI spec version: 0.1.0
 */

export interface ClusterSummary {
  cluster: number;
  label: string;
  count: number;
  avgImpactScore: number;
}