# lib/api-zod/src/generated/types/fullStatPlayer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSL AI Cricket Team Recommendation API
 * OpenAPI spec version: 0.1.0
 */

export interface FullStatPlayer {
  name: string;
  role: string;
  bowlingType?: string;
  matches: number;
  totalRuns: number;
  totalWickets: number;
  battingAvg: number;
  strikeRate: number;
  hundreds?: number;
  fifties?: number;
  ducks?: number;
  highestScore?: string;
  bowlingAvg?: number;
  economy?: number;
  bowlStrikeRate?: number;
  wicketsPerMatch?: number;
  finalScore: number;
  roleScore: number;
  performanceClass: string;
  clusterLabel?: string;
}