# lib/api-zod/src/generated/types/fullStatsResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSL AI Cricket Team Recommendation API
 * OpenAPI spec version: 0.1.0
 */
import type { FullStatPlayer } from "./fullStatPlayer";

export interface FullStatsResponse {
  players: FullStatPlayer[];
  total: number;
}