# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSL AI Cricket Team Recommendation API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  name: string;
  role: string;
  bowlingType?: string;
  impactScore: number;
  bayesianScore: number;
  finalScore: number;
  roleScore: number;
  cluster: string;
  clusterLabel: string;
  battingAvg?: number;
  strikeRate?: number;
  recentForm?: number;
  wicketsPerMatch?: number;
  economy?: number;
  matches: number;
  venueScore?: number;
  performanceClass: string;
  venueBattingAvg?: number;
  venueStrikeRate?: number;
  venueBowlingAvg?: number;
  venueBowlingSR?: number;
  venueWickets?: number;
  venueEconomy?: number;
}