# lib/api-zod/src/generated/types/recommendResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSL AI Cricket Team Recommendation API
 * OpenAPI spec version: 0.1.0
 */
import type { ModelInfo } from "./modelInfo";
import type { TeamBalance } from "./teamBalance";
import type { TeamPlayer } from "./teamPlayer";

export interface RecommendResponse {
  team: TeamPlayer[];
  venue: string;
  opposition?: string;
  teamBalance: TeamBalance;
  modelInfo: ModelInfo;
}