# lib/api-zod/src/generated/types/topPerformersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PSL AI Cricket Team Recommendation API
 * OpenAPI spec version: 0.1.0
 */
import type { Player } from "./player";

export interface TopPerformersResponse {
  batsmen: Player[];
  allRounders: Player[];
  fastBowlers: Player[];
  spinBowlers: Player[];
  wicketkeepers: Player[];
  venue?: string;
}